style: 格式化

+++ contract/contracts/Migrations.sol
@@ -2,23 +2,23 @@
 pragma solidity >=0.4.22 <0.9.0;
 
 contract Migrations {
-  address public owner;
+    address public owner;
 
-  uint public last_completed_migration;
+    uint256 public last_completed_migration;
 
-  constructor() {
-    owner = msg.sender;
-  }
+    constructor() {
+        owner = msg.sender;
+    }
 
-  modifier restricted() {
-    require(
-      msg.sender == owner,
-      "This function is restricted to the contract's owner"
-    );
-    _;
-  }
+    modifier restricted() {
+        require(
+            msg.sender == owner,
+            "This function is restricted to the contract's owner"
+        );
+        _;
+    }
 
-  function setCompleted(uint completed) public restricted {
-    last_completed_migration = completed;
-  }
+    function setCompleted(uint256 completed) public restricted {
+        last_completed_migration = completed;
+    }
 }
